create util for download link creation
